Move audit board sign-in page, form into submodule.
